global_server/requirements.txt
torch==2.2.1+cpu
torchvision==0.17.1+cpu
flwr==1.7.0
numpy
pyyaml
flwr_datasets
